+++ apps/docs/src/app/core/component-docs/radio/examples/radio-examples.component.ts
@@ -4,4 +4,11 @@ import { Component } from '@angular/core';
     selector: 'fd-radio-example',
     templateUrl: './radio-example.component.html'
 })
-export class RadioExamplesComponent {}
+export class RadioExamplesComponent {
+    optionVariable: string = 'val1';
+    optionTwoVariable: string = 'val2';
+    optionThreeVariable: string = 'val1';
+    optionFourVariable: string = 'val1';
+    optionFiveVariable: string = 'val1';
+    optionSixVariable: string = 'val1';
+}