+++ libs/core/src/lib/popover/popover.component.ts
@@ -40,6 +40,10 @@ export class PopoverComponent {
     @Input()
     noArrow: boolean = true;
 
+    /** Whether the popover container needs an extra class for styling. */
+    @Input()
+    addContainerClass: string;
+
     /** Whether the popover is disabled. */
     @Input()
     @HostBinding('class.fd-popover-custom--disabled')
@@ -113,7 +117,7 @@ export class PopoverComponent {
     /** Id of the popover. If none is provided, one will be generated. */
     @Input()
     id: string = 'fd-popover-' + popoverUniqueId++;
-
+    
     /**
      * Toggles the popover open state.
      */